+++ @utahdts/utah-design-system/react/components/forms/FormContextProvider.jsx
@@ -92,7 +92,7 @@ function FormContextProvider({
       },
       state,
     }),
-    [formState, onChange, onSubmit, setState, state]
+    [formState, state]
   );
 
   return (