+++ @utahdts/utah-design-system/react/components/table/Table.jsx
@@ -1,4 +1,6 @@
+// @ts-check
 import PropTypes from 'prop-types';
+import React from 'react';
 import RefShape from '../../propTypesShapes/RefShape';
 import joinClassNames from '../../util/joinClassNames';
 
@@ -14,15 +16,23 @@ const defaultProps = {
   id: null,
 };
 
+/**
+ * @param {Object} props
+ * @param {React.ReactNode} props.children
+ * @param {string | null} [props.className]
+ * @param {React.RefObject | null} [props.innerRef]
+ * @param {string | null} [props.id]
+ * @returns {JSX.Element}
+ */
 function Table({
   children,
-  className,
-  innerRef,
-  id,
+  className = null,
+  id = null,
+  innerRef = null,
   ...rest
 }) {
   return (
-    <table className={joinClassNames('table', className)} id={id} ref={innerRef} {...rest}>
+    <table className={joinClassNames('table', className)} id={id ?? undefined} ref={innerRef} {...rest}>
       {children}
     </table>
   );